LayoutTests/fast/forms/interactive-validation-lost-focusable-expected.txt

Test a case that the first invalid focusable control becomes invisible by a scroll event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The form was not submitted
PASS document.activeElement is document.getElementById("i3")
TEST COMPLETE
   
